+++ resources/release-notes.md
@@ -1,3 +1,7 @@
+## App
+
+- (All) Add "Health" tab to display health check results
+
 ## Software
 
-- (IMAP) Allow running IMAP analysis without any input data pattern
+- (IMAP) Add health checks for secondary currents and voltages, saturation, missed ITFs, and activation